double  => int

+++ Engine/Engine/ITileManager.cs
@@ -5,7 +5,7 @@ namespace Engine.Engine
     public interface ITileManager
     {
         void Add(SpriteOverlay sprite);
-        void AddBlockTile(double X, double Y, int Id, int size, bool SholdDraw);
+        void AddBlockTile(int X, int Y, int Id, int size, bool SholdDraw);
         void LoadMap(string MapData);
         void RemoveTile(Block tile);
     }
